requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pydantic
python-dotenv
requests
openai
tiktoken
weaviate-client
beautifulsoup4
httpx
twilio
pydantic-settings
python-multipart
pyyaml
websockets
pydub
